chore: rename `sunpy/data/manager` to `sunpy/data/data_manager`
